Show querystrings, query counts, IP etc.
Stash metadata at the start of each log file to record the URL path &
querystring, remote IP etc.

Display those in the list.  If the querystring is too long, truncate it
and provide an onclick to display the full thing - so the list of API
route calls can be neat enough to see, but the info is easily available.