Display call location when viewing queries
When displaying the queries made during a HTTP request, show the top
level frame of interest from the stack trace - skipping over DBIx::Class
etc internals that are unlikely to be of interest.

I will probably extend it to have a "click to view full stack trace"
feature, but this gets the bit that's most likely to be of use for now.