Make URL clickable, for GET requests at least
When displaying a HTTP request and the queries it ran, it makes sense
for the URL path to be a clickable link, if it was a GET request (it
doesn't make much sense if it was a PUT/POST, so don't bother there).

If we amend the profiling stuff to capture request parameters in full
for PUT/POST, then we *could* have a hidden form which re-does that -
but we'd want to avoid that for cases where it involved say a large file
upload - so probably simplest to only do this for GET for now.